feat: Move ibc workaround methods in ExternalStaking to interface
